requirements.txt
streamlit
streamlit-cookies-manager
numpy>=1.21.0,<1.24.0
pycryptodome>=3.19.0
pillow
tensorflow
librosa
torch
torchvision
timm
pandas
pyrebase
firebase-admin
opencv-python-headless
decord
tqdm
albumentations
requests==2.26.0
requests-toolbelt==0.9.1
